ALC100_captacion/serializers.py
from rest_framework import serializers

from ALC100_captacion.models.models import Convocatoria

class ConvocatoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Convocatoria
        fields = '__all__'  

class CandidatoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Convocatoria
        fields = '__all__'
